sliver appBar

+++ lib/Features/Cats/data/CatsModels/cats_image_model.dart
@@ -2,8 +2,8 @@ import 'package:equatable/equatable.dart';
 
 class Image extends Equatable {
   final String id;
-  final int width;
-  final int height;
+  final double width;
+  final double height;
   final String url;
 
   const Image({
@@ -16,8 +16,8 @@ class Image extends Equatable {
   factory Image.fromJson(Map<String, dynamic> json) {
     return Image(
       id: json['id'] ?? '',
-      width: json['width'] ?? 0,
-      height: json['height'] ?? 0,
+      width: (json['width'] ?? 0).toDouble(),
+      height: (json['height'] ?? 0).toDouble(),
       url: json['url'] ?? '',
     );
   }
